Indicate in confirm and consult messages the address that was responsible for the request.
Add "sessionuser" variable to a Majordomo object, indicating the
address that was used to initiate a session.

  $fh = new IO::File ">$file" || die "Eek!";
does not work as expected.  Use "or" instead of ||.

Add "posing" variable to access rules; it indicates if the
user for the current request is different from the original
"sessionuser."  The "posing" variable has side-effects:  if the rule
for a request is "mismatch" and "posing" is true, both
the user and victim may be consulted.  "posing" is treated
the same as "mismatch" for (un)subscribe commands.

Change the set and show commands to use "mismatch" by
default for access checks.

Enhance "default user" command: send results of each
command to the posed address.  The poser will still
see the result of the command (succeed, stall, fail)
but none of the output.  The message to the posee
has the poser's address in the From: header and
majordomo-owner in Reply-To.