(_post) Obtain message date from the last (earliest) Received header, or the Date header, or use the current time.
With this arrangement, messages sent in July but approved in August
will be stored in the July archive.  The advantage is that old
mail messages can be imported using the post-archive command.
The disadvantage is that external archiving software which uses
only the current month's archive may miss the older messages.